# lib/api-zod/src/generated/types/facultyInput.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * YMKCOE College Website API
 * OpenAPI spec version: 0.1.0
 */

export interface FacultyInput {
  title?: string;
  name: string;
  department: string;
  designation: string;
  qualification?: string;
  specialization?: string;
  experience?: string;
  researchGuideStatus?: string;
  email?: string;
  phone?: string;
  photoUrl?: string;
  adminRoles?: string;
  coreSkills?: string;
  publications?: string;
  professionalMemberships?: string;
  bio?: string;
  isHOD?: boolean;
}