# lib/api-zod/src/generated/types/placementInput.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * YMKCOE College Website API
 * OpenAPI spec version: 0.1.0
 */

export interface PlacementInput {
  studentName: string;
  company: string;
  package: string;
  role?: string;
  department: string;
  year: number;
  logoUrl?: string;
  testimonial?: string;
}